% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{meteroites.hist}
\alias{meteroites.hist}
\title{meteroites.hist}
\usage{
\code{meteroites.hist()}
}
\arguments{
\item{l}{defines the x limit of the included mass in the histogram. The limit starts from 0.}
}
\description{
\code{meteroites.hist} is a function for plotting a histogram of the meteroits depending on the mass.
}
