% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{meteroites.plotClass}
\alias{meteroites.plotClass}
\title{meteroites.plotClass}
\usage{
\code{meteroites.plotClass()}
WARNING: Before using this function, \code{\link{clusterComputation()}} must be executed.
}
\arguments{
\item{No}{additional argument must be used.}
}
\description{
\code{meteroites.plotClass} is a function for plotting a graph of the different classes of meteroites.
The x-axis contains the different classes. The y-axis contains the number of meteroites in each class.
}
