% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{meteroitesapi}
\alias{meteroitesapi}
\title{meteroitesapi}
\usage{
meteroitesapi()
}
\description{
meteroitesapi is a function to downlaod the meteroites data from the NASA SODA API.
It returns a data frame.
}
