% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{meteroites.heatMap}
\alias{meteroites.heatMap}
\title{meteroites.heatMap}
\usage{
meteroites.heatMap()
}
\description{
\code{meteroites.heatMap} is a function for plotting a heatmap of the relationship between the types of meteroites in the meteroitesData file and the class of meteroites .
}
\seealso{
\link{distance}
}
